% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{destinos_semana}
\alias{destinos_semana}
\title{Plotagem do numero de destinos de um aeroporto}
\usage{
destinos_semana(destinos = NULL, min_ocorencias, title)
}
\arguments{
\item{destinos}{vetor de string dos aeroportos de destino de interesse. Use a sigla FAA igual espeficicado no pacote \code{anyflyghts}. Default = NULL tras todos os destinos}

\item{min_ocorencias}{numero minimo de ocorrencias para que seja considerado um destino regular.}

\item{title}{titulo a ser utilizado pelo grafico}
}
\value{
um grafico
}
\description{
Essa função faz a plotagem do numero de destinos de um aeroporto, divididos por dia da semana e considerando apenas os destinos com um minimo de ocorrencias anual
}
\examples{
destinos_semana(destinos = c('SFO', 'LAX', 'LAS', 'BOS', 'IAD', 'ORD'),
                min_ocorencias = 100, title = "Destinos clave por dia da semana")
}
