% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{destinos_semana_tabela}
\alias{destinos_semana_tabela}
\title{Tabela de coincidencia origem-destino}
\usage{
destinos_semana_tabela(destinos = NULL, min_ocorencias = NULL)
}
\arguments{
\item{destinos}{vetor de string dos aeroportos de destino de interesse. Use a sigla FAA igual espeficicado no pacote \code{anyflyghts}. Default = NULL tras todos os destinos}

\item{min_ocorencias}{numero minimo de ocorrencias para que seja considerado um destino regular.}
}
\value{
uma tabela
}
\description{
Produz uma tabela com todas as origem nas linhas e destinos selecionados nas colunas, marcando 'X' se o voo existir.
}
\examples{
destinos_semana_tabela(destinos = c('SFO', 'LAX'), min_ocorencias = 100)
}
