% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{max_antecipado}
\alias{max_antecipado}
\title{Funcao auxiliar - encontra o voo de maior antecipacao de partida}
\usage{
max_antecipado(data)
}
\arguments{
\item{data}{dataset \code{meus_voos}}
}
\value{
uma linha do dataset \code{meus_voos}
}
\description{
Funcao auxiliar - encontra o voo de maior antecipacao de partida
}
\examples{
max_antecipado(meus_voos)
}
