% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{max_atraso}
\alias{max_atraso}
\title{Funcao auxiliar - encontra o voo de maior atraso de partida}
\usage{
max_atraso(data)
}
\arguments{
\item{data}{dataset \code{meus_voos}}
}
\value{
uma linha do dataset \code{meus_voos}
}
\description{
Funcao auxiliar - encontra o voo de maior atraso de partida
}
\examples{
max_atraso(meus_voos)
}
