% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_atraso}
\alias{plot_atraso}
\title{Funcao para plotagem do atraso percentual medio ao longo do ano}
\usage{
plot_atraso(col, title, atraso_min)
}
\arguments{
\item{col}{uma coluna do tipo \code{difftime} do dataset \code{meus_voos}}

\item{title}{um stirng com o titulo a ser usado pelo grafico}

\item{atraso_min}{declara o valor minimo em segundos para ser considerado um atraso. Use \code{lubridate::make_difftime()} para declarar o valor em segundos}
}
\value{
um grafico
}
\description{
Essa funcao calcula o atraso percentual medio de atrasos ao longo do ano, com base em um valor minino de segundos, para os 6 aeroportos da base de dados.
}
\examples{
plot_atraso(col = atraso_partida, title = 'partida > 30min',
            atraso_min = lubridate::make_difftime(60*30))
}
