% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_atraso_serie}
\alias{plot_atraso_serie}
\title{Funcao de plotagem do atraso percentual medio ao longo do tempo}
\usage{
plot_atraso_serie(atraso_min)
}
\arguments{
\item{atraso_min}{declara o valor minimo em segundos para ser considerado um atraso. Use \code{lubridate::make_difftime()} para declarar o valor em segundos.}
}
\value{
um grafico
}
\description{
Essa funcao faz a plotagem do atraso percentual medio ao longo os meses para os 6 aeroportos da base de dados.
}
\examples{
plot_atraso_serie(atraso_min = lubridate::make_difftime(60*30))
}
