% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_count}
\alias{plot_count}
\title{Contagem de categorias com base em uma coluna}
\usage{
plot_count(col, title)
}
\arguments{
\item{col}{uma coluna categorica do dataset \code{meus_voos}}

\item{title}{um stirng com o titulo a ser usado pelo grafico}
}
\value{
um grafico de barras com a contagem
}
\description{
Essa funcao conta a quantidade de nivels existe de uma coluna sempre agrupadas pela coluna \code{origem} do dataset \code{meus_voos}.
}
\examples{
plot_count(col = companhia , title = 'Numero de companhia aéreas operando voos')
}
